# specs/001-create-rustic-ai/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Redis Messaging Debugger API
  description: API for visualizing and debugging RusticAI message flow through Redis
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Development server
paths:
...
  /guilds/{guildId}/topics/{topicName}/messages:
    get:
      summary: Get messages for a specific topic
      operationId: getTopicMessages
      tags:
        - Messages
      parameters:
        - name: guildId
          in: path
          required: true
          schema:
            type: string
        - name: topicName
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: sinceId
          in: query
          schema:
            type: string
          description: GemstoneID to fetch messages after
        - name: minutes
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 10080  # 7 days
            default: 60
          description: Time window in minutes
      responses:
        '200':
          description: Paginated list of messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMessages'
  /messages/{messageId}:
    get:
      summary: Get a specific message by ID
      operationId: getMessage
      tags:
        - Messages
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
          description: GemstoneID of the message
      responses:
        '200':
          description: Message details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Message not found
...
components:
  schemas:
...
    Message:
      type: object
      required:
        - id
        - guildId
        - topicId
        - timestamp
        - type
        - payload
        - sender
        - routing
        - status
        - version
      properties:
        id:
          type: string
        guildId:
          type: string
        topicId:
          type: string
        threadId:
          type: string
        timestamp:
          type: string
          format: date-time
        type:
          type: string
        payload:
          type: object
        sender:
          $ref: '#/components/schemas/AgentIdentifier'
        routing:
          $ref: '#/components/schemas/RoutingSlip'
        status:
          $ref: '#/components/schemas/ProcessStatus'
        error:
          $ref: '#/components/schemas/ErrorInfo'
        ttl:
          type: integer
        tags:
          type: array
          items:
            type: string
        version:
          type: string
    AgentIdentifier:
      type: object
      required:
        - id
        - name
        - type
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
    ProcessStatus:
      type: string
      enum:
        - PENDING
        - PROCESSING
        - SUCCESS
        - ERROR
        - TIMEOUT
        - SKIPPED
        - RETRY
    RoutingSlip:
      type: object
      required:
        - entries
        - currentIndex
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/RoutingEntry'
        currentIndex:
          type: integer
    RoutingEntry:
      type: object
      required:
        - agentId
        - topicId
        - timestamp
        - status
      properties:
        agentId:
          type: string
        topicId:
          type: string
        timestamp:
          type: string
          format: date-time
        duration:
          type: integer
        status:
          $ref: '#/components/schemas/ProcessStatus'
        error:
          type: string
    ErrorInfo:
      type: object
      required:
        - code
        - message
        - timestamp
        - retryable
      properties:
        code:
          type: string
        message:
          type: string
        stack:
          type: string
        context:
          type: object
        timestamp:
          type: string
          format: date-time
        retryable:
          type: boolean
        retryCount:
          type: integer
...
    PaginatedMessages:
      type: object
      required:
        - data
        - pagination
        - timestamp
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        pagination:
          type: object
          required:
            - page
            - pageSize
            - total
            - hasNext
            - hasPrev
          properties:
            page:
              type: integer
            pageSize:
              type: integer
            total:
              type: integer
            hasNext:
              type: boolean
            hasPrev:
              type: boolean
        timestamp:
          type: string
          format: date-time
  securitySchemes:
    featureFlag:
      type: apiKey
      in: header
      name: X-Feature-Replay